requirements.txt
fastapi
uvicorn
pydantic
langchain-groq
pandas
streamlit
python-dotenv